add complete profile page, protect routes, fix bugs
Added complete profile page which will make a request to the
backend to create a local user based on the OAuth user that we have.
Added a api request function for converting OAuth user to Local user,
Protected some routes, fixed some Links to have replace set to true,
and more.